third_party/blink/web_tests/external/wpt/html/canvas/offscreen/text/2d.text.draw.baseline.hanging-expected.txt

This is a testharness.js-based test.
FAIL OffscreenCanvas test: 2d.text.draw.baseline.hanging assert_approx_equals: Red channel of the pixel at (5, 5) expected 0 +/- 2 but got 255
Harness: the test ran to completion.
